refactor: extract SidebarContainer from DocsLayout
